fix: setting currentIndex localStorage: issue when creating/deleting calendar fixed; CreateCalendarModal: re-factoring
