+++ fe-running-scheduler/src/components/CreateSchedule/index.js
@@ -1,0 +1,4 @@
+import CreateScheduleControls from "./CreateScheduleControls";
+import CreateScheduleForm from "./CreateScheduleForm";
+
+export { CreateScheduleControls, CreateScheduleForm };
